Update to fix the 3D- visualisation of VAT in Lead Macaque
I'm actually working with Andy on the new toolbox macaque to fix the error messages. And so, when we press stimulation in the 3D-visualisation window there is always a message says that there is a problem in load. So I tried to add if options to make it works in macaque Mode.