Set 'pad2' field in header when converting FTR to trk file.
This field stores the orientation code of the reference image of the
fiber coordinates.
If 'specs' paramter is not specified in 'ea_ftr2trk', the coordinates
are supposed to be in MNI space. So the reference image will be MNI T1
image. The 'pad2' field will be 'RAS '.
If the 'specs' parameter is already provided, 'pad2' will be calculated
based on 'specs.affine' using 'aff2axcodes'.