BUG: fix path format issues (for matlab tools and windows platform)
path has '"\path\to\file\T1w.nii"' structure.

however if using fileparts, that will trim to '"\path\to\file\' with the " termination, which will cause issues in the subsequent code.
This causes ea_mkdir to fail, or creation of invalid concatenated paths.

this is because cfg.moving, cfg.fixed, and cfg.outputimage are wrapped by quotes ("") when passed to ea_ants_run
this is caused by using ea_path_helper, which on windows adds the quotes to solve issues with spaces (which it is not done on unix/macOS/ubuntu)

added new method ea_path_formatlab, which is supposed to "invert" what is done by ea_path_helper, for a single path for now.

what was affected:
ea_mkdir(ea_fileparts(cfg.outputimage));
fixedinit, movinit, logDir