BUG: Fix indexing issue (crossval fit to score with subsets)
The issue was that when fitting to scores (doactualprediction), UseI was created without taking into account the patient selection, leading to issues when using a subset of patients and indexing this with the training and test vectors of cvp, that is build based on patient selection. Same issue with covariates.
Note that in some cases obj.patientselection should be taken into account, and in others it should be obj.customselection: for this reason we need to always index with patientsel