FIX: fix export for VIM patients
VIM export of localisation is broken, as the "VIM View" atlas is not present anymore. This fix uses the DISTAL Medium atlas, which is present in lead-dbs. Drawback is a slightly longer loading time during export, as it loads the whole DISTAL medium atlas before creating the views and .ply files